KAFKA-4801: don't verify assignment during broker up and down in testConsumptionWithBrokerFailures (#11949)
In this test, we have another thread to let broker down and up, to test if consumer can still work as expected. During the broker down and up, we tried to verify the assignment is as what we expected. But the rebalance will keep triggering while broker down and up. It doesn't make sense to verify the assignment here. Remove it to make the test reliable.

Reviewers: Guozhang Wang <wangguoz@gmail.com>